Combine image filter logic to imgutil
